Rename `assets.rs` to `helpers.rs`
